chore(experiments): moving the legacy experiment view to the product folder
