+++ products/experiments/frontend/legacy/components/LegacyDataCollection.tsx
@@ -15,9 +15,10 @@ import {
     CachedLegacyExperimentQueryResponse,
 } from '~/queries/schema/schema-general'
 import { DEFAULT_MDE } from '~/scenes/experiments/constants'
-import { legacyExperimentLogic, getInsightType } from '~/scenes/experiments/legacy'
 import { formatUnitByQuantity } from '~/scenes/experiments/utils'
 import { Experiment, InsightType } from '~/types'
+
+import { legacyExperimentLogic, getInsightType } from 'products/experiments/frontend/legacy'
 
 const getFirstPrimaryMetric = (experiment: Experiment): ExperimentTrendsQuery | ExperimentFunnelsQuery | null => {
     if (experiment.metrics.length) {
