+++ products/experiments/frontend/legacy/components/LegacyFunnelAggregationSelect.tsx
