+++ products/experiments/frontend/legacy/hooks/useLegacySvgResizeObserver.ts
