+++ products/experiments/frontend/legacy/legacyExperimentModalsLogic.ts
@@ -1,6 +1,6 @@
 import { MakeLogicType, actions, kea, path, reducers } from 'kea'
 
-import { SharedMetric } from '../SharedMetrics/sharedMetricLogic'
+import { SharedMetric } from '../../../../frontend/src/scenes/experiments/SharedMetrics/sharedMetricLogic'
 
 // Generated by kea-typegen. Update if you're an agent, ignore if you're human.
 export interface legacyExperimentModalsLogicValues {
