+++ products/experiments/frontend/legacy/metricsView/LegacyTickPanel.tsx
